requirements.txt
psycopg2-binary
slack-sdk
python-dotenv
requests